Added test for PersonnelMemberController

+++ src/main/java/com/dezoo/personeel/controllers/PersonnelController.java
@@ -6,9 +6,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.annotation.PostConstruct;
-import java.text.ParseException;
-import java.text.SimpleDateFormat;
 import java.util.List;
 
 @RestController
@@ -24,18 +21,20 @@ public class PersonnelController {
     }
 
     @GetMapping("/{personnelID}")
-    public PersonnelMember getPersonnelMemberID(@PathVariable String personnelID) {
-        return personelRepository.findPersonelMemberByPersonelId(personnelID);
+    public PersonnelMember getPersonnelMemberByID(@PathVariable String personnelID) {
+        return personelRepository.findPersonnelMemberByPersonnelId(personnelID);
     }
 
     @PostMapping
-    public void postPersonnelMember(@RequestBody PersonnelMember personnelMember) {
+    public ResponseEntity<PersonnelMember> postPersonnelMember(@RequestBody PersonnelMember personnelMember) {
         personelRepository.save(personnelMember);
+
+        return ResponseEntity.ok(personnelMember);
     }
 
-    @PutMapping("/{personnelID}")
-    public ResponseEntity<PersonnelMember> putPersonnelMember(@PathVariable String personnelID, @RequestBody PersonnelMember personnelMember) {
-        PersonnelMember personnelMember2 = personelRepository.findPersonelMemberByPersonelId(personnelID);
+    @PutMapping
+    public ResponseEntity<PersonnelMember> putPersonnelMember(@RequestBody PersonnelMember personnelMember) {
+        PersonnelMember personnelMember2 = personelRepository.findPersonnelMemberByPersonnelId(personnelMember.getPersonnelId());
 
         personnelMember2.setAddress(personnelMember.getAddress());
         personnelMember2.setDateOfBirth(personnelMember.getDateOfBirth());
@@ -45,27 +44,29 @@ public class PersonnelController {
         personnelMember2.setPrivatePhoneNumber(personnelMember.getPrivatePhoneNumber());
         personnelMember2.setPersonelCategory(personnelMember.getPersonelCategory());
 
-        personnelMember2 = personelRepository.save(personnelMember2);
+        personelRepository.save(personnelMember2);
 
         return ResponseEntity.ok(personnelMember2);
     }
 
     @DeleteMapping("/{personnelID}")
-    public ResponseEntity<PersonnelMember> deletePersonnelMember(@PathVariable String personnelID){
-        PersonnelMember personnelMember = personelRepository.findPersonelMemberByPersonelId(personnelID);
-        personelRepository.delete(personnelMember);
-
-        return ResponseEntity.ok(personnelMember);
+    public ResponseEntity<PersonnelMember> deletePersonnelMember(@PathVariable String personnelID) {
+        PersonnelMember personnelMember = personelRepository.findPersonnelMemberByPersonnelId(personnelID);
+        if (personnelMember != null) {
+            personelRepository.delete(personnelMember);
+            return ResponseEntity.ok().build();
+        } else
+            return ResponseEntity.notFound().build();
     }
 
-    @PostConstruct
-    public void fillDBwithTestData() {
-        SimpleDateFormat format = new SimpleDateFormat("dd/MM/yyyy");
-        try {
-            personelRepository.save(new PersonnelMember("fs161100", "Ferre", "Snyers", format.parse("16/11/2000"), "Gestelstraat 21", "2250", "+32441439", "Administration"));
-            personelRepository.save(new PersonnelMember("cn170999", "Christophe", "Neefs", format.parse("17/09/1999"), "Lier", "2500", "", "Rabbits"));
-            personelRepository.save(new PersonnelMember("rh031000", "Robbe", "Heremans", format.parse("03/10/2000"), "Westerloo", "2260", "", "Lions"));
-        } catch (ParseException ignored) {
-        }
-    }
+//    @PostConstruct
+//    public void fillDBwithTestData() {
+//        SimpleDateFormat format = new SimpleDateFormat("dd/MM/yyyy");
+//        try {
+//            personelRepository.save(new PersonnelMember("fs161100", "Ferre", "Snyers", format.parse("16/11/2000"), "Gestelstraat 21", "2250", "+32441439", "Administration"));
+//            personelRepository.save(new PersonnelMember("cn170999", "Christophe", "Neefs", format.parse("17/09/1999"), "Lier", "2500", "", "Rabbits"));
+//            personelRepository.save(new PersonnelMember("rh031000", "Robbe", "Heremans", format.parse("03/10/2000"), "Westerlo", "2260", "", "Lions"));
+//        } catch (ParseException ignored) {
+//        }
+//    }
 }
